+++ test/test_general.py
@@ -13,6 +13,10 @@ def test_nobackup(tmp_path: Path) -> None:
     p.write_text(TEXT)
     with InPlace(p) as fp:
         assert not fp.closed
+        files = pylistdir(tmp_path)
+        assert len(files) == 2
+        assert files[0].startswith("._in_place-")
+        assert files[1] == "file.txt"
         for line in fp:
             assert isinstance(line, str)
             fp.write(line.swapcase())
@@ -27,9 +31,16 @@ def test_backup_ext(tmp_path: Path) -> None:
     p = tmp_path / "file.txt"
     p.write_text(TEXT)
     with InPlace(p, backup_ext="~") as fp:
+        assert not fp.closed
+        files = pylistdir(tmp_path)
+        assert len(files) == 2
+        assert files[0].startswith("._in_place-")
+        assert files[1] == "file.txt"
         for line in fp:
             fp.write(line.swapcase())
-    assert pylistdir(tmp_path) == ["file.txt", "file.txt~"]
+        assert not fp.closed
+    assert fp.closed
+    assert pylistdir(tmp_path) == ["file.txt", "file.txt~"]  # type: ignore[unreachable]
     assert p.with_suffix(".txt~").read_text() == TEXT
     assert p.read_text() == TEXT.swapcase()
 
@@ -41,6 +52,10 @@ def test_backup(tmp_path: Path) -> None:
     bkp = tmp_path / "backup.txt"
     with InPlace(p, backup=bkp) as fp:
         assert not fp.closed
+        files = pylistdir(tmp_path)
+        assert len(files) == 2
+        assert files[0].startswith("._in_place-")
+        assert files[1] == "file.txt"
         for line in fp:
             fp.write(line.swapcase())
         assert not fp.closed
@@ -48,27 +63,6 @@ def test_backup(tmp_path: Path) -> None:
     assert pylistdir(tmp_path) == ["backup.txt", "file.txt"]  # type: ignore[unreachable]
     assert bkp.read_text() == TEXT
     assert p.read_text() == TEXT.swapcase()
-
-
-def test_backup_ext_and_backup(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    bkp = tmp_path / "backup.txt"
-    with pytest.raises(ValueError):
-        InPlace(p, backup=bkp, backup_ext="~")
-    assert pylistdir(tmp_path) == ["file.txt"]
-    assert p.read_text() == TEXT
-
-
-def test_empty_backup_ext(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    with pytest.raises(ValueError):
-        InPlace(p, backup_ext="")
-    assert pylistdir(tmp_path) == ["file.txt"]
-    assert p.read_text() == TEXT
 
 
 def test_error_backup_ext(tmp_path: Path) -> None:
@@ -121,12 +115,13 @@ def test_delete_backup(tmp_path: Path) -> None:
     p = tmp_path / "file.txt"
     p.write_text(TEXT)
     bkp = tmp_path / "backup.txt"
-    with pytest.raises(OSError):
-        with InPlace(p, backup=bkp) as fp:
-            for i, line in enumerate(fp):
-                fp.write(line.swapcase())
-                if i == 5:
-                    p.unlink()
+    with InPlace(p, backup=bkp) as fp:
+        for i, line in enumerate(fp):
+            fp.write(line.swapcase())
+            if i == 5:
+                p.unlink()
+        with pytest.raises(OSError):
+            fp.close()
     assert pylistdir(tmp_path) == []
 
 
@@ -138,20 +133,8 @@ def test_early_close_nobackup(tmp_path: Path) -> None:
         for line in fp:
             fp.write(line.swapcase())
         fp.close()
-    assert pylistdir(tmp_path) == ["file.txt"]
-    assert p.read_text() == TEXT.swapcase()
-
-
-def test_early_close_and_write_nobackup(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    with pytest.raises(ValueError):
-        with InPlace(p) as fp:
-            for line in fp:
-                fp.write(line.swapcase())
-            fp.close()
-            fp.write("And another thing...\n")
+        assert pylistdir(tmp_path) == ["file.txt"]
+        assert p.read_text() == TEXT.swapcase()
     assert pylistdir(tmp_path) == ["file.txt"]
     assert p.read_text() == TEXT.swapcase()
 
@@ -165,24 +148,25 @@ def test_early_close_backup(tmp_path: Path) -> None:
         for line in fp:
             fp.write(line.swapcase())
         fp.close()
+        assert pylistdir(tmp_path) == ["backup.txt", "file.txt"]
+        assert bkp.read_text() == TEXT
+        assert p.read_text() == TEXT.swapcase()
     assert pylistdir(tmp_path) == ["backup.txt", "file.txt"]
     assert bkp.read_text() == TEXT
     assert p.read_text() == TEXT.swapcase()
 
 
-def test_early_close_and_write_backup(tmp_path: Path) -> None:
+def test_late_close(tmp_path: Path) -> None:
     assert pylistdir(tmp_path) == []
     p = tmp_path / "file.txt"
     p.write_text(TEXT)
-    bkp = tmp_path / "backup.txt"
-    with pytest.raises(ValueError):
-        with InPlace(p, backup=bkp) as fp:
-            for line in fp:
-                fp.write(line.swapcase())
-            fp.close()
-            fp.write("And another thing...\n")
-    assert pylistdir(tmp_path) == ["backup.txt", "file.txt"]
-    assert bkp.read_text() == TEXT
+    with InPlace(p) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_text() == TEXT.swapcase()
+    fp.close()
+    assert pylistdir(tmp_path) == ["file.txt"]
     assert p.read_text() == TEXT.swapcase()
 
 
@@ -194,20 +178,8 @@ def test_rollback_nobackup(tmp_path: Path) -> None:
         for line in fp:
             fp.write(line.swapcase())
         fp.rollback()
-    assert pylistdir(tmp_path) == ["file.txt"]
-    assert p.read_text() == TEXT
-
-
-def test_rollback_and_write_nobackup(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    with pytest.raises(ValueError):
-        with InPlace(p) as fp:
-            for line in fp:
-                fp.write(line.swapcase())
-            fp.rollback()
-            fp.write("And another thing...\n")
+        assert pylistdir(tmp_path) == ["file.txt"]
+        assert p.read_text() == TEXT
     assert pylistdir(tmp_path) == ["file.txt"]
     assert p.read_text() == TEXT
 
@@ -221,21 +193,40 @@ def test_rollback_backup(tmp_path: Path) -> None:
         for line in fp:
             fp.write(line.swapcase())
         fp.rollback()
+        assert pylistdir(tmp_path) == ["file.txt"]
+        assert p.read_text() == TEXT
     assert pylistdir(tmp_path) == ["file.txt"]
     assert p.read_text() == TEXT
 
 
-def test_rollback_and_write_backup(tmp_path: Path) -> None:
+def test_rollback_then_inner_close(tmp_path: Path) -> None:
     assert pylistdir(tmp_path) == []
     p = tmp_path / "file.txt"
     p.write_text(TEXT)
-    bkp = tmp_path / "backup.txt"
-    with pytest.raises(ValueError):
-        with InPlace(p, backup=bkp) as fp:
-            for line in fp:
-                fp.write(line.swapcase())
-            fp.rollback()
-            fp.write("And another thing...\n")
+    with InPlace(p) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+        fp.rollback()
+        assert pylistdir(tmp_path) == ["file.txt"]
+        assert p.read_text() == TEXT
+        fp.close()
+        assert pylistdir(tmp_path) == ["file.txt"]
+        assert p.read_text() == TEXT
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_text() == TEXT
+
+
+def test_rollback_then_outer_close(tmp_path: Path) -> None:
+    assert pylistdir(tmp_path) == []
+    p = tmp_path / "file.txt"
+    p.write_text(TEXT)
+    with InPlace(p) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+        fp.rollback()
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_text() == TEXT
+    fp.close()
     assert pylistdir(tmp_path) == ["file.txt"]
     assert p.read_text() == TEXT
 
@@ -346,76 +337,6 @@ def test_different_dir_file_backup(
     assert p.read_text() == TEXT.swapcase()
 
 
-def test_backup_dirpath(tmp_path: Path) -> None:
-    """
-    Assert that using a path to a directory as the backup path raises an error
-    when closing
-    """
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    not_a_file = tmp_path / "not-a-file"
-    not_a_file.mkdir()
-    assert pylistdir(not_a_file) == []
-    fp = InPlace(p, backup=not_a_file)
-    fp.write("This will be discarded.\n")
-    with pytest.raises(OSError):
-        fp.close()
-    assert pylistdir(tmp_path) == ["file.txt", "not-a-file"]
-    assert p.read_text() == TEXT
-    assert pylistdir(not_a_file) == []
-
-
-def test_backup_nosuchdir(tmp_path: Path) -> None:
-    """
-    Assert that using a path to a file in a nonexistent directory as the backup
-    path raises an error when closing
-    """
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    fp = InPlace(p, backup=tmp_path / "nonexistent" / "backup.txt")
-    fp.write("This will be discarded.\n")
-    with pytest.raises(OSError):
-        fp.close()
-    assert pylistdir(tmp_path) == ["file.txt"]
-    assert p.read_text() == TEXT
-
-
-def test_nonexistent(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    with pytest.raises(FileNotFoundError):
-        InPlace(p)
-    assert pylistdir(tmp_path) == []
-
-
-def test_with_nonexistent(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    with pytest.raises(FileNotFoundError):
-        with InPlace(p):
-            raise AssertionError("Not reached")
-    assert pylistdir(tmp_path) == []
-
-
-def test_nonexistent_backup_ext(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    with pytest.raises(FileNotFoundError):
-        InPlace(p, backup_ext="~")
-    assert pylistdir(tmp_path) == []
-
-
-def test_with_nonexistent_backup_ext(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    with pytest.raises(FileNotFoundError):
-        with InPlace(p, backup_ext="~"):
-            raise AssertionError("Not reached")
-    assert pylistdir(tmp_path) == []
-
-
 def test_reentrant_backup_ext(tmp_path: Path) -> None:
     assert pylistdir(tmp_path) == []
     p = tmp_path / "file.txt"
@@ -443,38 +364,99 @@ def test_use_and_reenter_backup_ext(tmp_path: Path) -> None:
     assert p.read_text() == TEXT.swapcase()
 
 
-def test_useless_after_close(tmp_path: Path) -> None:
+def test_same_backup_path(tmp_path: Path) -> None:
     assert pylistdir(tmp_path) == []
     p = tmp_path / "file.txt"
     p.write_text(TEXT)
-    with InPlace(p, backup_ext="~") as fp:
-        assert not fp.closed
-    assert fp.closed
-    with pytest.raises(ValueError):  # type: ignore[unreachable]
-        fp.flush()
-    with pytest.raises(ValueError):
-        next(fp)
-    with pytest.raises(ValueError):
-        fp.read()
-    with pytest.raises(ValueError):
-        fp.readline()
-    with pytest.raises(ValueError):
-        fp.readlines()
-    with pytest.raises(ValueError):
-        fp.write("")
-    with pytest.raises(ValueError):
-        fp.writelines([""])
-
-
-def test_rollback_too_late(tmp_path: Path) -> None:
-    assert pylistdir(tmp_path) == []
-    p = tmp_path / "file.txt"
-    p.write_text(TEXT)
-    with InPlace(p, backup_ext="~") as fp:
+    with InPlace(p, backup=p) as fp:
         for line in fp:
             fp.write(line.swapcase())
-    with pytest.raises(ValueError):
-        fp.rollback()
-    assert pylistdir(tmp_path) == ["file.txt", "file.txt~"]
-    assert p.with_suffix(".txt~").read_text() == TEXT
+    assert pylistdir(tmp_path) == ["file.txt"]
     assert p.read_text() == TEXT.swapcase()
+
+
+@pytest.mark.skipif(
+    platform.system() == "Windows", reason="Windows barely has file modes"
+)
+def test_copy_executable_perm(tmp_path: Path) -> None:
+    p = tmp_path / "file.txt"
+    p.write_text(TEXT)
+    p.chmod(0o755)
+    with InPlace(p) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_text() == TEXT.swapcase()
+    assert p.stat().st_mode & 0o777 == 0o755
+
+
+def test_empty_newline(tmp_path: Path) -> None:
+    BYTES = (
+        b"'Twas brillig, and the slithy toves\n"
+        b"\tDid gyre and gimble in the wabe;\r"
+        b"All mimsy were the borogoves,\r\n"
+        b"\tAnd the mome raths outgrabe.\n"
+    )
+    p = tmp_path / "file.txt"
+    p.write_bytes(BYTES)
+    with InPlace(p, newline="") as fp:
+        lines = fp.readlines()
+        assert lines == [
+            "'Twas brillig, and the slithy toves\n",
+            "\tDid gyre and gimble in the wabe;\r",
+            "All mimsy were the borogoves,\r\n",
+            "\tAnd the mome raths outgrabe.\n",
+        ]
+        fp.writelines(ln.swapcase() for ln in lines)
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_bytes() == BYTES.swapcase()
+
+
+def test_unix_newline(tmp_path: Path) -> None:
+    BYTES = (
+        b"'Twas brillig, and the slithy toves\n"
+        b"\tDid gyre and gimble in the wabe;\r"
+        b"All mimsy were the borogoves,\r\n"
+        b"\tAnd the mome raths outgrabe.\n"
+    )
+    p = tmp_path / "file.txt"
+    p.write_bytes(BYTES)
+    with InPlace(p, newline="\n") as fp:
+        lines = fp.readlines()
+        assert lines == [
+            "'Twas brillig, and the slithy toves\n",
+            "\tDid gyre and gimble in the wabe;\rAll mimsy were the borogoves,\r\n",
+            "\tAnd the mome raths outgrabe.\n",
+        ]
+        fp.writelines(ln.swapcase() for ln in lines)
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_bytes() == BYTES.swapcase()
+
+
+def test_dos_newline(tmp_path: Path) -> None:
+    BYTES = (
+        b"'Twas brillig, and the slithy toves\n"
+        b"\tDid gyre and gimble in the wabe;\r"
+        b"All mimsy were the borogoves,\r\n"
+        b"\tAnd the mome raths outgrabe.\n"
+    )
+    p = tmp_path / "file.txt"
+    p.write_bytes(BYTES)
+    with InPlace(p, newline="\r\n") as fp:
+        lines = fp.readlines()
+        assert lines == [
+            (
+                "'Twas brillig, and the slithy toves\n"
+                "\tDid gyre and gimble in the wabe;\r"
+                "All mimsy were the borogoves,\r\n"
+            ),
+            "\tAnd the mome raths outgrabe.\n",
+        ]
+        fp.writelines(ln.swapcase() for ln in lines)
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert p.read_bytes() == (
+        b"'tWAS BRILLIG, AND THE SLITHY TOVES\r\n"
+        b"\tdID GYRE AND GIMBLE IN THE WABE;\r"
+        b"aLL MIMSY WERE THE BOROGOVES,\r\r\n"
+        b"\taND THE MOME RATHS OUTGRABE.\r\n"
+    )