+++ test/test_symlinks.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import os
 from os.path import relpath
 from pathlib import Path
@@ -85,3 +86,70 @@ def test_symlink_backup(tmp_path: Path) -> None:
     assert bkp.read_text() == TEXT
     assert link.read_text() == TEXT.swapcase()
     assert real.read_text() == TEXT.swapcase()
+
+
+def test_backup_is_symlink(tmp_path: Path) -> None:
+    p = tmp_path / "file.txt"
+    p.write_text(TEXT)
+    realdir = tmp_path / "realdir"
+    realdir.mkdir()
+    realfile = realdir / "realfile.txt"
+    realfile.write_text("This is a symlinked file.\n")
+    linkdir = tmp_path / "linkdir"
+    linkdir.mkdir()
+    linkfile = linkdir / "linkfile.txt"
+    target = relpath(realfile, linkdir)
+    linkfile.symlink_to(target)
+    with InPlace(p, backup=linkfile) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+    assert pylistdir(tmp_path) == ["file.txt", "linkdir", "realdir"]
+    assert not p.is_symlink()
+    assert p.read_text() == TEXT.swapcase()
+    assert pylistdir(realdir) == ["realfile.txt"]
+    assert not realfile.is_symlink()
+    assert realfile.read_text() == "This is a symlinked file.\n"
+    assert pylistdir(linkdir) == ["linkfile.txt"]
+    assert not linkfile.is_symlink()
+    assert linkfile.read_text() == TEXT
+
+
+def test_file_links_to_backup(tmp_path: Path) -> None:
+    p = tmp_path / "file.txt"
+    backup = tmp_path / "backup.txt"
+    backup.write_text(TEXT)
+    target = "backup.txt"
+    p.symlink_to(target)
+    with InPlace(p, backup=backup) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+    assert pylistdir(tmp_path) == ["backup.txt", "file.txt"]
+    assert p.is_symlink()
+    assert os.readlink(p) == target
+    assert backup.read_text() == TEXT.swapcase()
+
+
+def test_backup_links_to_file(tmp_path: Path) -> None:
+    p = tmp_path / "file.txt"
+    p.write_text(TEXT)
+    backup = tmp_path / "backup.txt"
+    target = "file.txt"
+    backup.symlink_to(target)
+    with InPlace(p, backup=backup) as fp:
+        for line in fp:
+            fp.write(line.swapcase())
+    assert pylistdir(tmp_path) == ["backup.txt", "file.txt"]
+    assert not p.is_symlink()
+    assert p.read_text() == TEXT.swapcase()
+    assert not backup.is_symlink()
+    assert backup.read_text()
+
+
+def test_broken_symlink(tmp_path: Path) -> None:
+    p = tmp_path / "file.txt"
+    target = "nowhere.txt"
+    p.symlink_to(target)
+    with pytest.raises(OSError):
+        InPlace(p)
+    assert pylistdir(tmp_path) == ["file.txt"]
+    assert os.readlink(p) == target